AtomicMC (onefile).
